requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-classic
langchain-groq
langgraph
chromadb
pdfplumber
python-dotenv
sentence-transformers
tavily-python
langchain-tavily
torch
rank_bm25
langchain-experimental
supabase
langfuse
langchain-unstructured
unstructured-client
langchain-nvidia-ai-endpoints
langchain-google-genai
llama-parse